hello world

+++ server.c
@@ -1,8 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   server.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: obastug <obastug@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/31 13:29:10 by obastug           #+#    #+#             */
+/*   Updated: 2024/12/31 13:37:10 by obastug          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <signal.h>
 #include <unistd.h>
 #include <stdlib.h>
 
-#include <stdio.h>
 void	ft_putchar(char c)
 {
 	write(1, &c, 1);
@@ -20,19 +31,19 @@ void	ft_putnbr(int nb)
 		ft_putnbr(nb / 10);
 		nb = nb % 10;
 	}
-	if (nb < 10) ft_putchar(nb + 48);
+	if (nb < 10)
+		ft_putchar(nb + 48);
 }
 
 void	ft_putstr(char *str)
 {
-    int i;
+	int	i;
 
-    while(str[i] != '\0')
-    {
-        write(1, &str[i], 1);
-        i++;
-    }
+	while (str[i])
+		i++;
+	write(STDOUT_FILENO, str, i);
 }
+
 void	listener(int sig)
 {
 	static int	status;
@@ -42,27 +53,20 @@ void	listener(int sig)
 	if (sig == SIGUSR1)
 		number = number + (1 << counter);
 	counter++;
-
-	if (status == 1)
+	if ((status == 1) && (counter == 32))
 	{
-		if (counter == 32)
-		{
-			kill(number, SIGUSR1);
-			counter = 0;
-			number = 0;
-			status = 0;
-		}
+		kill(number, SIGUSR1);
+		counter = 0;
+		number = 0;
+		status = 0;
 	}
-	else if (status == 0)
+	else if ((status == 0) && (counter == 8))
 	{
-		if (counter == 8)
-		{
-			if (number == 0)
-				status = 1;
-			ft_putchar(number);
-			counter = 0;
-			number = 0;
-		}
+		ft_putchar(number);
+		counter = 0;
+		if (number == 0)
+			status = 1;
+		number = 0;
 	}
 }
